+++ src/Function/FunctionInterface.php
@@ -4,4 +4,6 @@ namespace Pyncer\Database\Function;
 use Pyncer\Database\QueryStringInterface;
 
 interface FunctionInterface extends QueryStringInterface
-{}
+{
+    public function getTable(): string;
+}